Moved create data base to conftests and added log info for SQLALchemy errors
