+++ tests/test_rest_api.py
@@ -5,6 +5,7 @@ from f8a_worker.models import OSIORegisteredRepos
 from sqlalchemy import create_engine
 from f8a_worker.models import Base
 from pytest_mock import mocker
+from tests.conftest import create_database
 
 import requests
 import os
@@ -56,22 +57,3 @@ def test_register_api_endpoint(client, mocker):
     assert(jsn['data']["data"]["git_sha"] == payload["git_sha"])
     assert(jsn['data']["data"]["git_url"] == payload["git_url"])
     assert(jsn['data']["data"]["email_ids"] == payload["email_ids"])
-
-
-def create_database():
-    """Help method to create database."""
-    con_string = 'postgresql://{user}:{passwd}@{pg_host}:' + \
-                 '{pg_port}/{db}?sslmode=disable'
-    connection = con_string.format(
-        user=os.getenv('POSTGRESQL_USER'),
-        passwd=os.getenv('POSTGRESQL_PASSWORD'),
-        pg_host=os.getenv(
-            'PGBOUNCER_SERVICE_HOST',
-            'bayesian-pgbouncer'),
-        pg_port=os.getenv(
-            'PGBOUNCER_SERVICE_PORT',
-            '5432'),
-        db=os.getenv('POSTGRESQL_DATABASE'))
-    engine = create_engine(connection)
-    Base.metadata.drop_all(engine)
-    OSIORegisteredRepos.__table__.create(engine)
